requirements.txt
asgiref==3.6.0
astroid==2.13.2
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
distlib==0.3.6
Django==4.1.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
Flask==2.2.2
h11==0.14.0
idna==3.4
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
platformdirs==2.6.2
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
pylint==2.15.10
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
rest-framework-simplejwt==0.0.2
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tomlkit==0.11.6
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.17.1
webdriver-manager==3.8.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
